requirements.txt
streamlit>=1.25.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
pillow>=9.0.0
openai>=1.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-dotenv>=0.19.0
pandas>=1.3.0
opencv-python>=4.5.0
pytesseract>=0.3.8
